requirement.txt
django
djangorestframework
django-cors-headers
django-environ
mysqlclient